data-in-depth/restructure god components into smaller components and hooks
