+++ apps/main/app/features/scenarioExplorer/explorer/share/exportUtils.ts
@@ -1,8 +1,8 @@
 import JSZip from "jszip"
 import { themeValues } from "@repo/ui/themes/theme"
-import type { ShareItem } from "../../../../share/types"
-import { handlerForItem, type CsvLookups } from "../../../../share/variants"
-import { withExt } from "../../../../share/utils/filename"
+import type { ShareItem } from "./types"
+import { handlerForItem, type CsvLookups } from "./variants"
+import { withExt } from "./utils/filename"
 
 /**
  * Export utilities for Data Explorer
