+++ apps/main/app/features/scenarioExplorer/explorer/share/variants/barChart.ts
@@ -6,7 +6,7 @@
 
 import React, { useEffect } from "react"
 import type { ChartDataPoint } from "../../../../scenarios/components/shared/types"
-import { barChartDataToCSV } from "../../tools/panels/dataInDepth/utils/exportUtils"
+import { barChartDataToCSV } from "../exportUtils"
 import ShareScenarioCard from "../cards/ShareScenarioCard"
 import type { ShareItemOfType } from "../types"
 import type { VariantHandler } from "../variants"
