+++ apps/main/app/features/scenarioExplorer/explorer/tools/panels/dataInDepth/components/shared/PercentileMatrixSkeleton.tsx
