+++ apps/main/app/features/scenarioExplorer/explorer/tools/panels/dataInDepth/components/shared/SectionHeader.tsx
