+++ apps/main/app/features/scenarioExplorer/explorer/tools/panels/dataInDepth/components/shared/chartConstants.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared layout constants for Data in depth chart sections.
+ */
+
+/**
+ * Pixel size of the per-scenario tier glyph charts in the section headers
+ * (CWS, AG, reservoir storage). Roughly 1.5x the scenario-explorer tier size.
+ */
+export const TIER_CHART_SIZE = 90
