+++ apps/main/app/features/scenarioExplorer/explorer/tools/panels/dataInDepth/hooks/useMultiScenarioSlots.ts
@@ -26,18 +26,34 @@
  *     (s) => useSpillMonthly(s, "major"),
  *   )
  *
- * If we ever raise the data-explorer scenario cap above 12, bump
- * `MAX_FETCH_SLOTS`. The longer-term fix is to migrate each section
- * to consume `useBatchStatistics` (a single batched API call), at
- * which point per-scenario hook fan-out goes away entirely.
+ * This must stay in step with `MAX_DATA_IN_DEPTH_SCENARIOS`
+ * (`config/scenarioLimit.ts`), the cap on how many scenarios Data in Depth
+ * compares at once. If that cap changes, bump `MAX_FETCH_SLOTS` and add or
+ * remove the matching `fetchHook` calls below. The count has to be a literal
+ * here (one fetch slot per call) for the Rules of Hooks, so it cannot import
+ * the constant directly. The longer-term fix is to migrate each section to
+ * consume `useBatchStatistics` (a single batched API call), at which point the
+ * per-scenario hook fan-out goes away entirely.
  */
 
-const MAX_FETCH_SLOTS = 12
+const MAX_FETCH_SLOTS = 6
 
 export function useMultiScenarioSlots<T>(
   scenarios: string[],
   fetchHook: (scenarioId: string | null) => T,
 ): T[] {
+  if (
+    process.env.NODE_ENV !== "production" &&
+    scenarios.length > MAX_FETCH_SLOTS
+  ) {
+    // Selecting more scenarios than we have slots would silently drop the
+    // extras (they never get a fetch). Surface it loudly in development.
+    console.warn(
+      `useMultiScenarioSlots: ${scenarios.length} scenarios exceed MAX_FETCH_SLOTS (${MAX_FETCH_SLOTS}). ` +
+        `Scenarios beyond the cap will not load. Raise MAX_FETCH_SLOTS or migrate the section to useBatchStatistics.`,
+    )
+  }
+
   const slots: (string | null)[] = new Array(MAX_FETCH_SLOTS).fill(null)
   for (let i = 0; i < Math.min(scenarios.length, MAX_FETCH_SLOTS); i++) {
     slots[i] = scenarios[i] ?? null
@@ -51,13 +67,7 @@ export function useMultiScenarioSlots<T>(
   const r3 = fetchHook(slots[3] ?? null)
   const r4 = fetchHook(slots[4] ?? null)
   const r5 = fetchHook(slots[5] ?? null)
-  const r6 = fetchHook(slots[6] ?? null)
-  const r7 = fetchHook(slots[7] ?? null)
-  const r8 = fetchHook(slots[8] ?? null)
-  const r9 = fetchHook(slots[9] ?? null)
-  const r10 = fetchHook(slots[10] ?? null)
-  const r11 = fetchHook(slots[11] ?? null)
 
-  const all = [r0, r1, r2, r3, r4, r5, r6, r7, r8, r9, r10, r11]
+  const all = [r0, r1, r2, r3, r4, r5]
   return all.slice(0, scenarios.length)
 }
